requirements.txt
# 核心依赖
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
chromadb>=0.4.22  # 更新版本以支持Python 3.9+
pillow==10.1.0

# PDF处理
pdfplumber==0.10.3
PyPDF2==3.0.1
PyMuPDF>=1.23.0  # 使用较新版本以支持Apple Silicon

# OCR引擎
paddleocr>=2.8.0  # 更新版本,减少依赖冲突
paddlepaddle==2.6.2  # PaddleOCR依赖 (更新到可用版本)

# 图像处理
opencv-python>=4.6.0  # OpenCV图像处理

# CLIP模型
open-clip-torch==2.23.0

# 进度条和美化输出
tqdm==4.66.1
rich==13.7.0

# 工具库
numpy==1.24.3
regex==2023.10.3

# 测试（可选）
pytest==7.4.3
